% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_required_packages.R
\name{install_required_packages}
\alias{install_required_packages}
\title{Install required packages}
\usage{
install_required_packages(input_file)
}
\arguments{
\item{input_file}{Path to the R Markdown file}
}
\value{
A character vector with the names of the packages used
}
\description{
Installs or updates the packages used in an R Markdown file.
}
