% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_md.R
\name{rmd_to_md}
\alias{rmd_to_md}
\title{Convert external R Markdown to Markdown}
\usage{
rmd_to_md(rmd_file, output_dir)
}
\arguments{
\item{rmd_file}{Path to the R Markdown file. This can be a local path or a URL.}

\item{output_dir}{The output directory where the Markdown file and figures
will be saved.}
}
\value{
Markdown file written do disk.
}
\description{
This function converts an R Markdown (.Rmd) file to a Markdown (.md) file. It
can handle both local and remote .Rmd files.
}
\examples{
rmd_file <- "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main/vignettes/articles/occurrence-process.Rmd"

rmd_to_md(rmd_file, output_dir = "r/gcube")

# Clean up (don't do this if you want to keep your files)
unlink("r/gcube", recursive = TRUE)
}
