% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_frontmatter.R
\name{update_frontmatter}
\alias{update_frontmatter}
\title{Update front matter}
\usage{
update_frontmatter(md_file_path, rmd_file, title = NULL, sidebar_order = NULL)
}
\arguments{
\item{md_file_path}{Path to the markdown file on disk.}

\item{rmd_file}{Path to the R Markdown file, either a local path or a URL.}

\item{title}{Title of the article, to show in the menu.}

\item{sidebar_order}{Number indicating the order of the article in the
sidebar.}
}
\value{
Markdown file with updated front matter, written to disk.
}
\description{
Updates the front matter of a markdown file on disk.
}
\examples{
\dontrun{
md_file_path <- here::here(
"output", "src", "content", "docs", "software", "gcube",
"occurrence-process.md"
)
rmd_file <- "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main/vignettes/articles/occurrence-process.Rmd"
title <- "2. Occurrence process"
sidebar_order <- 2
update_frontmatter(md_file_path, rmd_file, title, sidebar_order)
}
}
