% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_md.R
\name{rmd_to_md}
\alias{rmd_to_md}
\title{Convert external R Markdown to Markdown}
\usage{
rmd_to_md(rmd_file, md_dir, fig_dir, fig_url_dir)
}
\arguments{
\item{rmd_file}{Path to the R Markdown file, either a local path or a URL.}

\item{md_dir}{Path to local directory to Markdown file to.}

\item{fig_dir}{Path to local directory to save figures to.}

\item{fig_url_dir}{URL path that will be used to link to the figures in the
markdown output.}
}
\value{
Markdown file and figures written do disk.
}
\description{
Converts an R Markdown (.Rmd) file to a Markdown (.md) file. Both local and
remote .Rmd files can be handled. The date this function is called upon is
added to the beginning of the Markdown file.
}
\examples{
rmd_file <- "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main/vignettes/articles/occurrence-process.Rmd"

md_dir <- file.path("output", "src", "content", "docs", "r", "gcube")
fig_dir <- file.path("output", "public", "r", "gcube")
fig_url_dir <- paste0("/astro-docs/", "r", "gcube", "/")
rmd_to_md(rmd_file, md_dir, fig_dir, fig_url_dir)

# Clean up (don't do this if you want to keep your files)
unlink("output", recursive = TRUE)
}
