% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_md.R
\name{rmd_to_md}
\alias{rmd_to_md}
\title{Convert external R Markdown to Markdown}
\usage{
rmd_to_md(
  rmd_file,
  md_dir,
  fig_dir,
  fig_url_dir,
  title = NULL,
  sidebar_label = NULL,
  sidebar_order = NULL
)
}
\arguments{
\item{rmd_file}{Path to the R Markdown file, either a local path or a URL.}

\item{md_dir}{Path to local directory to Markdown file to. If it doesn't
exist it will be created.}

\item{fig_dir}{Path to local directory to save figures to.}

\item{fig_url_dir}{URL path that will be used to link to the figures in the
markdown output.}

\item{title}{Title of the article, to show on top of the page}

\item{sidebar_label}{Title in the sidebar.}

\item{sidebar_order}{Number indicating the order of the article in the
sidebar.}
}
\value{
Markdown file and figures written do disk.
}
\description{
Converts an R Markdown (.Rmd) file to a Markdown (.md) file. Both local and
remote .Rmd files can be handled. The date this function is called upon is
added to the beginning of the Markdown file.
}
\examples{
\dontrun{
rmd_file <- "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main/vignettes/articles/occurrence-process.Rmd"

md_dir <- file.path("output", "src", "content", "docs", "software", "gcube")
fig_dir <- file.path("output", "public", "software", "gcube")
fig_url_dir <- "/software/gcube/"
title <- "2. Occurrence process"
sidebar_label <- "occurrence-process"
sidebar_order <- 2

# Don't forget to install (and update) required packages loaded in the Rmd file
# Convert Rmd to md
rmd_to_md(rmd_file, md_dir, fig_dir, fig_url_dir, title, sidebar_label, sidebar_order)

# Clean up (don't do this if you want to keep your files)
unlink("output", recursive = TRUE)
}
}
