% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_frontmatter.R
\name{update_frontmatter}
\alias{update_frontmatter}
\title{Update front matter}
\usage{
update_frontmatter(
  md_file_path,
  rmd_file,
  title = NULL,
  sidebar_label = NULL,
  sidebar_order = NULL,
  replace = NULL
)
}
\arguments{
\item{md_file_path}{Path to the Markdown file on disk.}

\item{rmd_file}{Path to the R Markdown file, either a local path or a URL.}

\item{title}{Title of the article, to show on top of the page.}

\item{sidebar_label}{Title in the sidebar.}

\item{sidebar_order}{Number indicating the order of the article in the
sidebar.}

\item{replace}{Set \code{key = value} pairs to replace all \code{key} strings by their
\code{value}.}
}
\value{
Markdown file with updated front matter, written to disk.
}
\description{
Updates the front matter and replaces the logo of a Markdown file on disk.
}
\examples{
\dontrun{
update_frontmatter(
  md_file_path = file.path(
    "output/src/content/docs/software/gcube/occurrence-process.md"
  ),
  rmd_file = file.path(
    "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main",
    "vignettes/articles/occurrence-process.Rmd"
  ),
  title = "2. Occurrence process",
  sidebar_label = "Occurrence process",
  sidebar_order = 2;
  replace = c("### Changing number of occurrences over time" =
              "### How to change the number of occurrences over time")
)
}
}
