% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_frontmatter.R
\name{update_frontmatter}
\alias{update_frontmatter}
\title{Update front matter}
\usage{
update_frontmatter(
  md_file,
  rmd_file,
  title = NULL,
  sidebar_label = NULL,
  sidebar_order = NULL,
  replace = NULL
)
}
\arguments{
\item{md_file}{Path to the Markdown file on disk.}

\item{rmd_file}{Path to the R Markdown file, either a local path or a URL.}

\item{title}{Title of the article, to show on top of the page.}

\item{sidebar_label}{Title in the sidebar.}

\item{sidebar_order}{Number indicating the order of the article in the
sidebar.}

\item{replace}{Named character vector with \code{c("key" = "value")} pairs.
All occurrences of \code{"key"} in the content of the Markdown file will be
replaced by their respective \code{"value"} (before updating the front matter).}
}
\value{
Markdown file with updated front matter (and optionally content),
written to disk.
}
\description{
Updates the front matter (and optionally content) of a Markdown file on disk.
}
\examples{
\dontrun{
update_frontmatter(
  md_file = file.path(
    "output/src/content/docs/software/gcube/occurrence-process.md"
  ),
  rmd_file = file.path(
    "https://raw.githubusercontent.com/b-cubed-eu/gcube/refs/heads/main",
    "vignettes/articles/occurrence-process.Rmd"
  ),
  title = "2. Occurrence process",
  sidebar_label = "Occurrence process",
  sidebar_order = 2,
  replace = c(
    "### Changing number of occurrences over time" =
    "### How to change the number of occurrences over time",
    "man/figures/logo.png" = "https://b-cubed-eu.github.io/gcube/logo.png"
  )
)
}
}
